+++ src/nodes/dwf.cpp
@@ -82,7 +82,7 @@ void DWF::deinit() {
 
 void DWF::setOut() {
 	vector<double> current_dwf = dwf_work.front().second;
-	for (int i = 0; i < Flow::size(); i++) {
+	for (size_t i = 0; i < Flow::size(); i++) {
 		out[i] = current_dwf[i];
 	}
 	out[0] = out[0] * q_scale;
